refactor(writes): move LockedIterator to write_through, make public
LockedIterator (was the private _LockedGen in remote_table_exec) only
needed threading + a generator -- it had no dependency on
remote_table_exec's contents. Living there forced relations.py into a
function-level import to dodge the relations <-> remote_table_exec module
cycle, and split the two advance-lock siblings across modules.

Move it next to DrainingIterator in writes/write_through.py:
- xorq.writes is cycle-free; both relations and remote_table_exec already
  import from it at module level, so the function-level workaround is gone.
- co-locates the two 'serialize next() vs teardown' classes (#2107).

Rename _LockedGen -> LockedIterator and drop the underscore: it is now a
cross-module export (in xorq.writes __all__), composed-adjacent to
DrainingIterator, and used in tests -- the private name was a soft lie.

test_tee_write_iter_guard_by_drain now patches the name in both the
relations and remote_table_exec namespaces, since each binds it at import.
