+++ python/xorq/expr/remote_table_exec.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import functools
-import threading
 import weakref
 from collections import Counter
 from typing import Any, Callable, NamedTuple
@@ -17,7 +16,7 @@ from xorq.expr.relations import RemoteTable, gen_name
 from xorq.vendor.ibis import Expr
 from xorq.vendor.ibis.backends import BaseBackend
 from xorq.vendor.ibis.expr import operations as ops
-from xorq.writes import DrainingIterator
+from xorq.writes import DrainingIterator, LockedIterator
 
 
 logger = get_logger(__name__)
@@ -282,57 +281,6 @@ class RemoteTableScope:
         self.close()
 
 
-class _LockedGen:
-    """Serializes advancement and close of a wrapped generator.
-
-    A datafusion scan can park the foreground ``next()`` inside an FFI pull
-    (the generator frame is in the *executing* state, not suspended) while a
-    background pull lags. If the ``weakref.finalize`` backstop then calls
-    ``close()`` on that frame, CPython raises ``ValueError: generator already
-    executing`` -- in a finalizer/daemon-thread context that escapes and aborts
-    the process. The shared lock guarantees ``close()`` only runs while the
-    generator is suspended, never mid-advance. Same race class as the
-    ``DrainingIterator`` advance lock (#2107).
-    """
-
-    def __init__(self, gen) -> None:
-        self._gen = gen
-        self._lock = threading.Lock()
-        self._closed = False
-
-    def __iter__(self) -> "_LockedGen":
-        return self
-
-    def __next__(self):
-        with self._lock:
-            if self._closed:
-                raise StopIteration
-            try:
-                return next(self._gen)
-            except StopIteration:
-                # natural exhaustion: mark closed so a later close() is a no-op
-                # and _closed tracks "no longer advanceable", not just explicit
-                # close calls.
-                self._closed = True
-                raise
-
-    def close(self) -> None:
-        # Non-blocking acquire: closing an *executing* frame raises
-        # ``ValueError: generator already executing``, and *blocking* until the
-        # advance returns can wedge the finalizer during interpreter teardown
-        # (GIL-release fatal). So if an advance holds the lock, skip the gen
-        # close entirely -- that in-flight consumer owns the generator's
-        # teardown (its ``finally`` runs ``scope.close()``); the caller's
-        # explicit ``scope.close()`` still fires regardless.
-        if self._lock.acquire(blocking=False):
-            try:
-                if not self._closed:
-                    self._closed = True
-                    self._gen.close()
-            finally:
-                self._lock.release()
-
-
 def bind_scope_to_reader(
     scope: RemoteTableScope,
     reader: pa.RecordBatchReader,
@@ -348,9 +296,9 @@ def bind_scope_to_reader(
     pyarrow 21 ``reader.close()`` alone does not finalize the wrapped
     generator -- cleanup then fires when the last reference drops.
 
-    The generator is wrapped in ``_LockedGen`` so the finalizer's ``close()``
+    The generator is wrapped in ``LockedIterator`` so the finalizer's ``close()``
     cannot race a foreground advance still parked inside a datafusion FFI pull
-    (see ``_LockedGen``).
+    (see ``LockedIterator``).
     """
 
     def gen():
@@ -364,7 +312,7 @@ def bind_scope_to_reader(
             # (remote_table_scope) surface drain failures instead.
             scope.close()
 
-    locked = _LockedGen(gen())
+    locked = LockedIterator(gen())
     out = pa.RecordBatchReader.from_batches(reader.schema, locked)
 
     def _cleanup() -> None:
